york indoors news, publish script fix for mac
git-svn-id: https://svn2.assembla.com/svn/yorkultimate@17 d6e9c351-0237-4953-a62f-a7a1ee97849f

+++ htdocs/publish.sh
@@ -25,7 +25,9 @@ for folder in $FOLDERS ; do
 	cp -rf "./$folder" "$dest""/""$folder";
 done;
 
-for svnfolder in `find -iregex "^[.\\/]*pub[\\/].+\.svn$"` ; do
-	echo "Deleting ""$svnfolder";
-	rm -rf $svnfolder;
-done;
+if [ $TERM_PROGRAM != "Apple_Terminal" ] ; then
+	for svnfolder in `find -iregex "^[.\\/]*pub[\\/].+\.svn$"` ; do
+		echo "Deleting ""$svnfolder";
+		rm -rf $svnfolder;
+	done;
+fi;
